# backend/packages/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AAAgronegócio Platform API - Farm Listings Marketplace & SaaS
 * OpenAPI spec version: 1.0.0
 */
import type { UserProfileRole } from "./userProfileRole";

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
  phone?: string | null;
  avatarUrl?: string | null;
  isPremium: boolean;
  isActive: boolean;
  createdAt: Date;
}